# api_rest/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto DAW Backend API",
        "description": "API REST del Proyecto DAW. Esta API gestiona la lógica de negocio para la plataforma de intercambio de tiempo.\n *     \n *     Módulos incluidos:\n *     - **Autenticación**: Registro, login, logout y gestión de perfil.\n *     - **Usuarios**: Gestión completa de usuarios (CRUD).\n *     - **Servicios**: Gestión de ofertas y demandas de servicios.\n *     - **Transacciones**: Control de intercambios de tiempo entre usuarios.\n *     - **Valoraciones**: Sistema de reputación y feedback.\n *     - **Mensajes**: Sistema de mensajería interna.\n *     - **Common**: Recursos compartidos (provincias, poblaciones, etc.).",
        "contact": {
            "email": "soporte@proyectodaw.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo local"
        },
        {
            "url": "https://proyecto-daw-backend.onrender.com/api",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener un usuario específico",
                "operationId": "b253b58cc9966940360ec90555bfe3ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario encontrado"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Pedro"
                                                },
                                                "apellido": {
                                                    "type": "string",
                                                    "example": "Garcia"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "pedro.garcia@example.com"
                                                },
                                                "provincia_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "ciudad_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "descripcion": {
                                                    "type": "string",
                                                    "example": "Descripción del usuario"
                                                },
                                                "horas_saldo": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "valoracion": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "rol_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario no Encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/users/{user}/change-password": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Cambiar contraseña del usuario",
                "operationId": "b86fcb6de14591c8ecbf34d6b3d15b1e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "nueva_contraseña123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contraseña actualizada"
                    },
                    "400": {
                        "description": "La contraseña es la misma"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Autenticación mediante token Laravel Sanctum. Ingrese el token directamente.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}